Start integrating battle mechanics with main program
Most of this progress is in the Battle class.

Also finished writing DamageMove class. Next steps include writing
classes for other types of moves, such as moves that drain enemy
health like Absorb and healing moves like Recover.

InfoField now includes a method that allows its components to
update based on the health of the Pokemon.

Renamed various getters in Pokemon class for conciseness: i.e.
getName() --> name(). Also added methods that provide info to the
main program, like the percentage of current health to max health
for a Pokemon.