Write basic game loop structure
Summary:
The player with the faster Pokemon has their moves go first after
both the player and the rival locks in their moves. For example,
if both the player and rival choose to attack, the player's attack
will be inflicted first. The same logic happens to switches.
Switches are given priority over attacks, so if a slow player switches
and the faster player attacks, the switch will happen first.