Add text for when stat stages are raised / lowered
This includes a mini rework as to how moves are used. Before, the use()
method would return an int. I needed a way to check if a move successfully
raised / lowered a stat stage. It would be natural for use() to return a
boolean in this case, but the return type had already been reserved as int
so I can use the damage values for other calculations. I made a compromise
using Object[] return type to incorporate both int and boolean types in one.
This way I can return false if the move was not successful and still have
int for damage values.

Also fixed a bug regarding Pokemon.setStatus. Before, a Pokemon's status
would 100% be set successfully. That's not how the games rules work, though.
For example, if a Pokemon is a Poison type, it's impossible to give them the
Poison status. I added in some checks to ensure these rules hold.