Remove assertion statements involving damage calculations
I realized that the damage range assertion doesn't take into account
the possibility of a critical hit, which multiplies the damage by a
factor of 1.5. As a result, if there is a critical hit, the test cases
will fail.

I will need to reconsider how to test for this. In the meantime, these
assertions have been removed.